Moved TxTreeGraph out of TxTree.[cpp,h] into include/klee/util/TxTreeGraph.h and lib/Core/TxTreeGraph.cpp
